refactor IQM2 scraper to mixin
